validation

+++ middlewares/task.js
@@ -1,7 +1,8 @@
 const fs = require('fs');
 const date = require("../helpers/getDate");
+const verify = require("../helpers/verifyKeys");
 const file = "./data/task.json";
-
+const time = "./data/time_record.json";
 
 
 module.exports = {
@@ -37,12 +38,16 @@ module.exports = {
       fs.readFile(file, 'utf8', (error,data)=>{
           if(error) return reject({ details: "error cargando datos", error });
           const obj = JSON.parse(data);
-          const newProject = {"id_task": parseInt(date.getDate()),"id_time_record": params.id_time_record ,"name": params.name, "date": params.date ,"description": params.description};
-          obj.push(newProject);
-          fs.writeFile(file, JSON.stringify(obj), function(err) {
-              if(err) return console.log(err);
-              return reply({details: "Saved correctly"});
-          }); 
+          verify.task_record(time,params.id_time_record, (result) =>{
+            console.log(result);
+              if(!result) return reply({details: "Error, id_time_record doesn´t exist"});
+              const newProject = {"id_task": parseInt(date.getDate()),"id_time_record": parseInt(params.id_time_record) ,"name": params.name, "date": params.date ,"description": params.description};
+              obj.push(newProject);
+              fs.writeFile(file, JSON.stringify(obj), function(err) {
+                  if(err) return console.log(err);
+                  return reply({details: "Saved correctly"});
+              }); 
+          });
       });
     })
   },
@@ -52,26 +57,29 @@ module.exports = {
        fs.readFile(file, 'utf8', (error,data)=>{
           if(error) return reject({ details: "error cargando datos", error });
           const obj = JSON.parse(data);
-          for(var i in obj){
-            if(obj[i].id_task == id_task){
-              obj[i].id_task = parseInt(id_task);
-              obj[i].id_time_record = parseInt(params.id_time_record);
-              obj[i].name = params.name;
-              obj[i].date = params.date;
-              obj[i].description = params.description;
-              bandera = true;
-            }
-          }
-          if(bandera){
-              fs.writeFile(file, JSON.stringify(obj), function(error) {
-                if(error) return console.log(error);
-                return reply({details: "Updated correctly"});
-              }); 
-          }else{
-             return reply({details: "the record doesnt exist,nothing to update"});           
-          }
+          verify.task_record(time,params.id_time_record, (result) =>{
+              console.log(result);
+            if(!rusult)return reply({details: "Error, id_time_record doesn´t exist"});
+               for(var i in obj){
+                  if(obj[i].id_task == id_task){
+                    obj[i].id_task = parseInt(id_task);
+                    obj[i].id_time_record = parseInt(params.id_time_record);
+                    obj[i].name = params.name;
+                    obj[i].date = params.date;
+                    obj[i].description = params.description;
+                    bandera = true;
+                  }
+                }
+                if(bandera){
+                    fs.writeFile(file, JSON.stringify(obj), function(error) {
+                      if(error) return console.log(error);
+                      return reply({details: "Updated correctly"});
+                    }); 
+                }else{
+                   return reply({details: "the record doesnt exist,nothing to update"});           
+                }
+          });
       });
-
     })
   },
   delete: (id_task) => {
